// src/interface.ts
export type AESMode = 'CBC' | 'CFB' | 'CTR' | 'OFB' | 'ECB'

/* Type aliases used to increase flexibility and be able
   to extend these types later on. Also type aliases allows
   names to be more self explanatory like in BASE58 case. */

export type TBytes = Uint8Array

export type TBase64 = string
export type TBase58 = string
export type TBase16 = string //Same as HEX

export type TChainId = string | number

//Every binary parameter could be represented as Uint8Array or number[] or base58 string
export type TBinaryIn = TBytes | TBase58 | number[]

export type TRawStringInDiscriminator = { TRawStringIn: null }

//Every input stinrg could be represented as Uint8Array or number[] or a string itself
export type TRawStringIn = TBytes | string | number[] | TRawStringInDiscriminator